AACForm/requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.0
bleach==3.1.4
cachetools==3.1.1
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2020.4.5.1
cffi==1.12.3
chardet==3.0.4
cssselect2==0.2.2
cycler==0.10.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.7
django-appconf==1.0.3
django-bootstrap4==1.1.1
django-extensions==2.2.3
django-filter==2.2.0
git+git://github.com/beknightedresearch/django-googledrive-storage@master
django-nose==1.4.6
django-summernote==0.8.11.6
django-weasyprint==0.5.5
djangorestframework==3.10.3
docutils==0.15.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
gunicorn==20.0.4
html5lib==1.0.1
httplib2==0.18.0
idna==2.9
Jinja2==2.10.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.2.0
model-bakery==1.0.2
nose==1.3.7
numpy==1.17.3
oauthlib==3.1.0
pandas==0.25.3
parsimonious==0.7.0
Pillow==7.0.0
psycopg2==2.8.3
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
pydot==1.4.1
pyflakes==2.1.1
pyparsing==2.4.2
PyPDF2==1.26.0
Pyphen==0.9.5
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.2
requests==2.23.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
tinycss2==1.0.2
uritemplate==3.0.0
urllib3==1.25.9
WeasyPrint==51
webencodings==0.5.1
whitenoise==5.0.1